Added a plan, made it work...
Added a plan that uses the module to setup the box, but also copies up fog, tmux.conf and .ssh files so that the account is more usable.

Had to split the zsh stuff out.. keeping it in the same class didn't work.  For some reason even though the zsh class required the user... the user creation failed because it was trying to set the shell to zsh... which it shouldn't do.
the module still works, but reuiqres running twice to work... which is lame.