fix for selinux issue
selinux is using a fact created by stdlib to get the vardir
When run from bolt it gets a bad value... bolt does one run to collect facts, and one to run the code...
and the var dir comes out as a tmp dir specific to each run.
So the vardir from the fact colection doesn't exist during the run.  Thus an error.
Probably wouldn't work if it did exist since it is temporary anyway.
So hardcoded it by passing a param into selinx