(SLV-379) Add Puppetfile
This commit adds a `Puppetfile` to define the modules and their dependencies
for installing this module via the `bolt puppetfile install` command.  This
is included because the module is not yet published to the Forge and cannot
be installed via the `puppet module install` command.

The following module dependency incompatibilities are known.

* The 'chocolatey-chocolatey' module depends on 'badgerious-windows_env'
  which has been deprecated.  The 'puppet-windows_env' module, which replaces
  the badgerious one, has been included

* The dependencies for 'puppetlabs-stdlib' cannot be satisfied for all of the
  modules involved.  Version '4.25.1' has been pinned in order meet the
  majority of the dependency requirements asked of it.