(SLV-435) Install bolt on host via setup plan
This commit adds the installation of puppet-bolt on the target
host using the `setup_instance` plan.  Since `apply_prep` is used
to install puppet previously, the necessary package repo is already
available on the host.  All that is needed is to ensure that the
package is installed via `puppet apply`.