(SLV-791) update for new dashboards
new grafana module for folder support
new task to setup the metrics collector repo on the runner
Switched to installing metrics collector instead of metrics viewer as the js2on2graphite was removed from the viewer and replaced as json2timeseriesdb in the collector.
Left the task for installing the viewer, as it would be valuable for testing the whole metrics ecosystem.  But left if off by default since we don't do that often.